+++ src/com/scurab/java/ftpleechergui/model/FTPStorageTableModel.java
@@ -158,7 +158,7 @@ public class FTPStorageTableModel extends AbstractTableModel {
             return name;
         } else if (columnIndex == 2) {
             if (f != null && f.isFile()) {
-                return TextUtils.getNumberReadable((int) f.getSize());
+                return TextUtils.getNumberReadable(f.getSize());
             }
         }
         return null;
